feat(mindmap): add judge when paste text, only select one node can paste text as node #WIK-11064
