+++ .changeset/yellow-crabs-cover.md
@@ -1,0 +1,5 @@
+---
+'@plait/mindmap': patch
+---
+
+add judge, only select one node can paste text as node
