+++ packages/mindmap/src/plugins/with-mindmap.ts
@@ -3,7 +3,6 @@ import {
     ELEMENT_TO_PLUGIN_COMPONENT,
     getSelectedElements,
     hotkeys,
-    idCreator,
     IS_TEXT_EDITABLE,
     Path,
     PlaitBoard,
@@ -27,7 +26,7 @@ import { getRectangleByNode, hitMindmapNode } from '../utils/graph';
 import { isVirtualKey } from '../utils/is-virtual-key';
 import { MINDMAP_ELEMENT_TO_COMPONENT } from '../utils/weak-maps';
 import { withNodeDnd } from './with-dnd';
-import { buildClipboardData, getDataFromClipboard, insertClipboardData, setClipboardData } from '../utils/clipboard';
+import { buildClipboardData, getDataFromClipboard, insertClipboardData, insertClipboardText, setClipboardData } from '../utils/clipboard';
 
 export const withMindmap: PlaitPlugin = (board: PlaitBoard) => {
     const { drawElement, dblclick, keydown, insertFragment, setFragment, deleteFragment, isIntersectionSelection } = board;
@@ -191,24 +190,7 @@ export const withMindmap: PlaitPlugin = (board: PlaitBoard) => {
         } else {
             const text = data?.getData(`text/plain`) as string;
             const textWidth = getWidthByText(text, board.host.parentElement as any);
-            if (text) {
-                const newElement = {
-                    id: idCreator(),
-                    value: {
-                        children: [{ text }]
-                    },
-                    children: [],
-                    width: textWidth,
-                    height: 24
-                };
-                const element = getSelectedElements(board)?.[0];
-                const nodeComponent = MINDMAP_ELEMENT_TO_COMPONENT.get(element as MindmapNodeElement);
-                if (nodeComponent) {
-                    const path = findPath(board, nodeComponent.node).concat(nodeComponent.node.children.length);
-                    Transforms.insertNode(board, newElement, path);
-                    return;
-                }
-            }
+            insertClipboardText(board, text, textWidth);
         }
         insertFragment(data, targetPoint);
     };
